# backend/cmd/docs/swagger.yaml
basePath: /
definitions:
  api.StatusBadRequestResponse:
    properties:
      error:
        type: string
    type: object
  api.StatusOkTagAddResponse:
    properties:
      added:
        items:
          $ref: '#/definitions/repositories.ImageTagSuccess'
        type: array
      errors:
        items:
          $ref: '#/definitions/repositories.ImageTagFail'
        type: array
    type: object
  api.StatusOkTagDeleteResponse:
    properties:
      deleted:
        items:
          $ref: '#/definitions/repositories.ImageTagSuccess'
        type: array
      errors:
        items:
          $ref: '#/definitions/repositories.ImageTagFail'
        type: array
    type: object
  api.StatusOkUploadResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/repositories.ImageUploadFail'
        type: array
      uploaded:
        items:
          $ref: '#/definitions/repositories.ImageUploadSuccess'
        type: array
    type: object
  api.StatusServerErrorResponse:
    properties:
      error:
        type: string
    type: object
  repositories.ImageTagFail:
    properties:
      id:
        type: integer
      reason:
        type: string
    type: object
  repositories.ImageTagSuccess:
    properties:
      count:
        type: integer
      id:
        type: integer
    type: object
  repositories.ImageUploadFail:
    properties:
      filename:
        type: string
      reason:
        type: string
    type: object
  repositories.ImageUploadSuccess:
    properties:
      filename:
        type: string
      id:
        type: integer
    type: object
info:
  contact: {}
  description: This is the API access for the backend of the FreezeTag app.
  title: FreezeTag API
  version: "0.1"
paths:
  /file/{id}:
    get:
      description: Get an image file given an ID
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: thumbnail file data
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.StatusBadRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.StatusServerErrorResponse'
      summary: Get file
  /search:
    get:
      description: Search for an image given information
      parameters:
      - description: camera make
        in: query
        name: make
        type: string
      - description: camera make fuzzy
        in: query
        name: makeLike
        type: string
      - description: camera model
        in: query
        name: model
        type: string
      - description: camera model fuzzy
        in: query
        name: modelLike
        type: string
      - description: picture taken before (unix epoch)
        in: query
        name: takenBefore
        type: string
      - description: picture taken after (unix epoch)
        in: query
        name: takenAfter
        type: string
      - description: picture uploaded before (unix epoch)
        in: query
        name: uploadedBefore
        type: string
      - description: picture uploaded after (unix epoch)
        in: query
        name: uploadedAfter
        type: string
      - description: latitude/longitude/distance (degrees)
        example: 100.0,12.0,1.0
        in: query
        name: near
        type: string
      - collectionFormat: multi
        description: picture tag
        in: query
        items:
          type: string
        name: tag
        type: array
      - collectionFormat: multi
        description: picture tag fuzzy
        in: query
        items:
          type: string
        name: tagLike
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              format: int64
              type: integer
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.StatusBadRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.StatusServerErrorResponse'
      summary: Search images
  /tag/add:
    post:
      description: Add tags to images
      parameters:
      - collectionFormat: multi
        description: tags to add
        in: query
        items:
          type: string
        name: tag
        required: true
        type: array
      - collectionFormat: multi
        description: image IDs to add to
        in: query
        items:
          type: integer
        name: id
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusOkTagAddResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.StatusBadRequestResponse'
      summary: Add tags
  /tag/list:
    get:
      description: Get all the tags in the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.StatusServerErrorResponse'
      summary: List all tags
  /tag/list/{id}:
    get:
      description: Get all the tags associated with an image
      parameters:
      - description: image ID to get the tags of
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.StatusBadRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.StatusServerErrorResponse'
      summary: List image tags
  /tag/remove:
    delete:
      description: Delete tags from images
      parameters:
      - collectionFormat: multi
        description: tags to remove
        in: query
        items:
          type: string
        name: tag
        required: true
        type: array
      - collectionFormat: multi
        description: image IDs to remove from
        in: query
        items:
          type: integer
        name: id
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusOkTagDeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.StatusBadRequestResponse'
      summary: Delete tags
  /thumbnails/{id}:
    get:
      description: Get a WEBP format image thumbnail for an image
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - image/webp
      responses:
        "200":
          description: thumbnail file data
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.StatusBadRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.StatusServerErrorResponse'
      summary: Get thumbnail
  /upload:
    post:
      description: Upload a set of image files to the server
      parameters:
      - collectionFormat: multi
        description: image file to upload
        in: formData
        items:
          type: file
        name: file
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusOkUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.StatusBadRequestResponse'
      summary: Upload files
swagger: "2.0"